+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - 调整docker安装
 - 新增k3sx卸载k3s
+- 新增dockerconfig配置docker
 
 ## 2022013015